+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - allow defining directives for interface types and theirs fields, with inheritance for object types fields (#744)
 ### Fixes
 - allow defining extension on field resolver level for fields also defined as a property of the class (#776)
+- fix throwing error when schema with dynamic default value was built again (#787)
 ### Others
 - **Breaking Change**: update `graphql-js` peer dependency to `^15.4.0`